requirements.txt
aiofiles==0.7.0
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.2
fastapi==0.65.1
fastapi-csv==0.1.0
greenlet==1.1.0
h11==0.12.0
idna==3.2
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
passlib==1.7.4
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydantic==1.8.2
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.17
starlette==0.14.2
typer==0.3.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
